Fixing no rows bug in TsQueryResponse
If the query is create with streaming set to false, no rows were in the
TsQueryResponse.Rows. The code defines a local variable and puts
the rows in it, and then the rows are lost. I removed the local variable
and set the rows direclty to get it working.

Add contributor